
\name{MatBase}
\alias{MatBase}
\title{Repeat for 2 vectors perpendicular placed}
\usage{
MatBase(x=c(1,3),y=1:10)
}
\description{
Repeat for 2 vectors perpendicular placed. For use in plotting orn completing the linkage to make one x,y,x coordinates for each point on the plane
}
\examples{
MatBase(x=c(1,3),y=1:10)
}
