\name{MatRand}
\alias{MatRand}
\title{Generate Matrix of Random Numbers}
\usage{
MatRand(x,y,min,max)
}
\description{
Generate Matrix of Random Numbers(usually for program testing).

Input vector x and y, the length of x and y will be used for the number of rows and columns.

The random number will range from min to max.
}
\examples{
Mat(2,5,min=-1,max=1)
}